refactor(papyrus_protobuf): move converters to rpc_transaction
